# api/openapi/fleet-api.yaml
openapi: "3.1.0"
info:
  title: fleet-llm-d Fleet Controller API
  version: "2.0.0"
  description: |
    REST API for the fleet-llm-d fleet controller. Manages multi-cluster GPU
    inference pools, tenant usage tracking, model rollouts, and ledger-based
    compliance verification. Fleet actions use the asynchronous v2 intent and
    operation lifecycle; admission is not evidence that an action executed.
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: fleet-llm-d maintainers
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/clusters:
    get:
      operationId: listClusters
      summary: List all registered clusters
      tags: [clusters]
      responses:
        "200":
          description: Array of cluster information objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClusterInfo"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: registerCluster
      summary: Register a new cluster
      tags: [clusters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterRegistration"
      responses:
        "201":
          description: Cluster registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/clusters/{id}:
    delete:
      operationId: deregisterCluster
      summary: Deregister a cluster
      tags: [clusters]
      parameters:
        - $ref: "#/components/parameters/ClusterID"
      responses:
        "204":
          description: Cluster deregistered successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/clusters/{id}/drain:
    post:
      operationId: drainCluster
      summary: Initiate graceful drain on a cluster
      tags: [clusters]
      parameters:
        - $ref: "#/components/parameters/ClusterID"
      responses:
        "200":
          description: Drain initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [draining]
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cluster is already draining or drained
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/clusters/{id}/activate:
    post:
      operationId: activateCluster
      summary: Reactivate a drained or draining cluster
      tags: [clusters]
      parameters:
        - $ref: "#/components/parameters/ClusterID"
      responses:
        "200":
          description: Cluster reactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [running]
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- pools ----------
  /api/v1/pools:
    get:
      operationId: listPools
      summary: List all fleet inference pools
      tags: [pools]
      responses:
        "200":
          description: Array of inference pool objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PoolInfo"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/pools/{name}/state:
    get:
      operationId: getPoolState
      summary: Get reconciled state for a single pool
      tags: [pools]
      parameters:
        - $ref: "#/components/parameters/PoolName"
      responses:
        "200":
          description: Reconciled pool state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolState"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- tenants ----------
  /api/v1/tenants:
    get:
      operationId: listTenants
      summary: List all tenants
      tags: [tenants]
      responses:
        "200":
          description: Array of tenant objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantInfo"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createTenant
      summary: Create a new tenant with quotas
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                name:
                  type: string
                priority:
                  type: integer
                quotas:
                  type: object
                rate_limit:
                  type: object
                cost_control:
                  type: object
      responses:
        "200":
          description: Tenant already exists (idempotent return)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInfo"
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Tenant already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tenants/{id}/usage:
    get:
      operationId: getTenantUsage
      summary: Get usage for a specific tenant
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/TenantID"
      responses:
        "200":
          description: Tenant usage data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUsage"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- metrics ----------
...
  /api/v1/agent/status:
    post:
      operationId: reportAgentStatus
      summary: Upsert a cluster status report from a fleet agent
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentStatusReport"
      responses:
        "200":
          description: Existing cluster status updated
        "201":
          description: New cluster status registered
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/agent/events:
    post:
      operationId: reportAgentEvent
      summary: Publish a structured event from a fleet agent
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentEventReport"
      responses:
        "202":
          description: Event accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/agent/policies/{cluster_id}:
    get:
      operationId: getAgentPolicies
      summary: Retrieve tenant quotas and placement constraints for a cluster
      tags: [agents]
      parameters:
        - name: cluster_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy sync response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicySyncResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- rollouts ----------
  /api/v1/rollouts:
    get:
      operationId: listRollouts
      summary: List all rollouts
      tags: [rollouts]
      responses:
        "200":
          description: Array of rollout objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RolloutInfo"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createRollout
      summary: Create a new rollout
      tags: [rollouts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolloutCreateRequest"
      responses:
        "201":
          description: Rollout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolloutInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/rollouts/{id}/promote:
    post:
      operationId: promoteRollout
      summary: Promote a canary rollout
      tags: [rollouts]
      parameters:
        - $ref: "#/components/parameters/RolloutID"
      responses:
        "200":
          description: Rollout promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolloutState"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/rollouts/{id}/rollback:
    post:
      operationId: rollbackRollout
      summary: Rollback a rollout
      tags: [rollouts]
      parameters:
        - $ref: "#/components/parameters/RolloutID"
      responses:
        "200":
          description: Rollout rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolloutState"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- cost ----------
  /api/v1/cost/pricing:
    get:
      operationId: getCostPricing
      summary: Get the full GPU pricing table
      tags: [cost]
      responses:
        "200":
          description: Array of GPU pricing entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GPUPricing"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/cost/tokenomics/{model}:
    get:
      operationId: getCostTokenomics
      summary: Compute token costs for a model across all GPU types
      tags: [cost]
      parameters:
        - $ref: "#/components/parameters/ModelName"
        - name: throughput
          in: query
          description: Tokens per second throughput (defaults to 1000)
          schema:
            type: number
            format: double
      responses:
        "200":
          description: Array of token cost entries per GPU type and tier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenCost"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/cost/chargeback/{tenant}:
    get:
      operationId: getCostChargeback
      summary: Generate a chargeback report for a tenant
      tags: [cost]
      parameters:
        - $ref: "#/components/parameters/TenantPathName"
        - name: budget
          in: query
          description: Monthly budget override (defaults to 10000)
          schema:
            type: number
            format: double
      responses:
        "200":
          description: Tenant chargeback report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargebackReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/cost/projection:
    get:
      operationId: getCostProjection
      summary: Project monthly cost based on current usage rates
      tags: [cost]
      parameters:
        - name: tokens_per_day
          in: query
          description: Estimated daily token consumption (defaults to 10000000)
          schema:
            type: integer
            format: int64
        - name: gpu_type
          in: query
          description: GPU type (defaults to H200)
          schema:
            type: string
        - name: tier
          in: query
          description: Pricing tier (defaults to on-demand)
          schema:
            type: string
        - name: model
          in: query
          description: Model name (defaults to default)
          schema:
            type: string
      responses:
        "200":
          description: Monthly cost projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostProjection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/cost/savings:
    get:
      operationId: getCostSavings
      summary: Compare current versus optimized placement cost
      tags: [cost]
      parameters:
        - name: current
          in: query
          description: Current monthly cost (defaults to 5000)
          schema:
            type: number
            format: double
        - name: optimized
          in: query
          description: Optimized monthly cost (defaults to 3000)
          schema:
            type: number
            format: double
      responses:
        "200":
          description: Savings projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsProjection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/cost/alerts:
    get:
      operationId: getCostAlerts
      summary: Check all tenant budgets and return active alerts
      tags: [cost]
      responses:
        "200":
          description: Array of active budget alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetAlert"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- governed intents and operations ----------
  /api/v1/intents:
    post:
      operationId: submitIntentV1Compatibility
      summary: Submit an intent through the deprecated v1 compatibility adapter
      deprecated: true
      description: |
        Compatibility adapter for clients migrating to `/api/v2/intents`.
        The response status `deferred` covers every non-terminal operation,
        including an intent that passed admission. `executed` is returned only
        after the associated operation reached `SUCCEEDED`, which requires an
        observed outcome and durable ledger evidence. Policy acceptance alone
        is never reported as execution.
      tags: [intents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FleetIntentRequest"
      responses:
        "200":
          description: Deprecated compatibility projection of the operation state
          headers:
            Deprecation:
              description: Always `true`; callers should migrate to v2.
              schema:
                type: string
                const: "true"
            Link:
              description: Link to the successor-version intent endpoint.
              schema:
                type: string
              example: '</api/v2/intents>; rel=successor-version'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntentResponseV1"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/intents:
    post:
      operationId: submitIntentV2
      summary: Admit a governed fleet intent asynchronously
      description: |
        Creates a durable intent and operation. A `202` response means the
        request was admitted to the lifecycle; it does not mean infrastructure
        changed. Follow `status_url` until a terminal operation state is
        reached. The same `idempotency_key` returns the original operation.
        Production ingress uses a verified GCL DecisionPackage CloudEvent.
        Plain JSON is unsigned, self-asserted operator compatibility and is
        rejected unless the server explicitly enables that development mode.
      tags: [intents]
      requestBody:
        required: true
        content:
          application/json:
            x-development-only: true
            schema:
              $ref: "#/components/schemas/OperatorCompatibilityIntentRequest"
          application/cloudevents+json:
            schema:
              $ref: "#/components/schemas/GCLDecisionPackageCloudEvent"
      responses:
        "202":
          description: Intent durably admitted or deferred for later processing
          headers:
            Location:
              description: Relative URL of the associated operation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: Invalid governed proposal, or intent reached `REJECTED` or `EXPIRED` during admission
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Error"
                  - $ref: "#/components/schemas/SubmissionResponse"
        "415":
          description: Request media type is unsupported, including plain JSON when operator compatibility is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: A GCL CloudEvent was submitted but trusted signature verification is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/intents/{id}:
    get:
      operationId: getIntentV2
      summary: Get an admitted fleet intent
      tags: [intents]
      parameters:
        - $ref: "#/components/parameters/IntentID"
      responses:
        "200":
          description: Immutable submitted intent and its evidence references
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FleetIntent"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/operations/{id}:
    get:
      operationId: getOperationV2
      summary: Get the durable lifecycle of a fleet operation
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationID"
      responses:
        "200":
          description: Operation state and append-only transition journal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FleetOperation"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/operations/{id}/approve:
    post:
      operationId: approveOperationV2
      summary: Record human approval for a deferred operation
      description: |
        Records an authenticated approval transition. Fleet admission and
        approval policy owns execution authorization. Immutable-ledger proof
        receipts record decisions but never grant authority.
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationID"
      responses:
        "200":
          description: Operation after recording the approval transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FleetOperation"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/operations/{id}/cancel:
    post:
      operationId: cancelOperationV2
      summary: Cancel or compensate a fleet operation
      description: |
        Before actuation, cancellation moves the operation to `FAILED`. During
        or after actuation it starts `ROLLING_BACK`; cancellation never claims
        that compensation completed synchronously.
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationID"
      responses:
        "200":
          description: Operation after recording cancellation or rollback intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FleetOperation"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- compliance ----------
  /api/v1/verify/chains:
    get:
      operationId: verifyChains
      summary: Verify all ledger decision chains
      description: |
        Validates the integrity of every decision chain stored in the ledger.
        Returns a map keyed by chain name to its verification result.
      tags: [compliance]
      responses:
        "200":
          description: Map of chain name to verification result
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/ChainVerification"
              example:
                placement:
                  valid: true
                  entries_checked: 142
                  chain_type: placement
                  verified_at: "2026-07-06T12:00:00Z"
                routing:
                  valid: true
                  entries_checked: 89
                  chain_type: routing
                  verified_at: "2026-07-06T12:00:00Z"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- auth ----------
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh an authentication token
      tags: [auth]
      responses:
        "200":
          description: New token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "401":
          description: Valid token required for refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- modelplane ----------
  /api/v1/modelplane/clusters:
    get:
      operationId: listModelPlaneClusters
      summary: List ModelPlane inference clusters
      tags: [modelplane]
      responses:
        "200":
          description: Array of ModelPlane cluster objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelPlaneCluster"
        "503":
          description: ModelPlane integration not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/modelplane/deployments:
    get:
      operationId: listModelPlaneDeployments
      summary: List ModelPlane model deployments
      tags: [modelplane]
      responses:
        "200":
          description: Array of ModelPlane deployment objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelPlaneDeployment"
        "503":
          description: ModelPlane integration not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/modelplane/cost/{deployment}:
    get:
      operationId: getModelPlaneDeploymentCost
      summary: Get the hourly cost of a ModelPlane deployment
      tags: [modelplane]
      parameters:
        - $ref: "#/components/parameters/DeploymentName"
      responses:
        "200":
          description: Deployment cost breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentCost"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: ModelPlane integration not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- webhook ----------
  /api/v1/webhook/fleetinferencepool:
    post:
      operationId: watchFleetInferencePool
      summary: Receive FleetInferencePool CRD watch events
      tags: [webhook]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchEvent"
      responses:
        "200":
          description: Watch event processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional Bearer token for authenticated access

  # ---------- reusable parameters ----------
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "name is required"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "internal server error"

  # ---------- schemas ----------
  schemas:
...
    ClusterInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        region:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
          description: Latest lifecycle phase reported by the cluster agent
        gpu_capacity:
          $ref: "#/components/schemas/GPUCapacity"
        utilization:
          type: number
          format: double
          description: Current cluster utilization (0.0 - 1.0)
      example:
        id: cluster-us-east-1a
        name: us-east-prod
        region: us-east-1
        labels:
          env: production
          tier: gpu
        status: Running
        gpu_capacity:
          available: 6
          total: 8
          types: ["A100"]
        utilization: 0.75
...